Creating New Objects from Existing Ones Part 1
Simply assigning a variable the value of an object will just make it point to the original one; a shallow copy.
By using Object.assign, it copies the fields over, making the top-level fields are independent, but anything nested will be both pointing to the same object.
There is a way to create fully deep copies of objects however, shown in the next part.