Lists and shoddy method to produce unique keys
A possible way to produce unique keys for sibling list options is to associate an index with each li tag; we can expand the map function to accommodate this; however, using the index is not a robust solution.

if the list can be modified by the user; this exposes the fragility of using index as keys, able to even modify incorrect tags, and produce very difficult-to-track issues

Better solution in next part!